requirements.txt
PyPDF2==3.0.1
langchain==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv==1.0.0
pytest==7.4.3 